fix: fixed font bug and created new component Login.tsx
