+++ tailwind.config.ts
@@ -13,7 +13,7 @@ const config: Config = {
 				"gradient-conic": "conic-gradient(from 180deg at 50% 50%, var(--tw-gradient-stops))",
 			},
 			fontFamily: {
-				roboto: ["Roboto", "sans-serif"],
+				sans: ["var(--font-inter)"],
 			},
 		},
 	},
